Day 1, Part 1
With O(1) insert and O(1) contains, we can leverage the HashSet to avoid quadratic performance.

Each number has exactly one other number that will add to 2020, namely 2020 - n.

Thus, we can compute what this "counterpart" would be, and check if we've seen it before,
and if so, return the product.